# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Job Manager Microservice.",
        "title": "Swagger Job Manager API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.4.1-latest"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
...
        "/jobmanager/jobs": {
            "get": {
                "description": "get all jobs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "List all Jobs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.Job"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Can not find Jobs",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "update a job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Update a Job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job UUID",
                        "name": "job_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Job information",
                        "name": "Job",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Job"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Job"
                        }
                    },
                    "400": {
                        "description": "Error reading request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Could not find job",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/jobmanager/jobs/executable/{orchestrator}/{owner_id}": {
            "get": {
                "description": "get jobs to execute",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "List Jobs to Execute",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Orchestrator type [ocm | nuvla]",
                        "name": "orchestrator",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Owner ID",
                        "name": "owner_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.Job"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Orchestrator type is required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Cannot find executable Jobs",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/jobmanager/jobs/promote/{job_uuid}": {
            "patch": {
                "description": "promote job by uuid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Promote Job by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job UUID",
                        "name": "job_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Job Promoted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Job UUID is required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Can not find Job to promote",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/jobmanager/jobs/{job_uuid}": {
            "get": {
                "description": "get job by uuid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get Job by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job UUID",
                        "name": "job_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/models.Job"
                        }
                    },
                    "400": {
                        "description": "Job UUID is required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Can not find Job by UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete job by uuid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Delete Job by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job UUID",
                        "name": "job_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Job UUID is required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Can not find Job to delete",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/jobmanager/resources/status": {
            "put": {
                "description": "update resource status by uuid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Update resource status by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resource UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Resource info",
                        "name": "resource",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Resource"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource updated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Resource UUID is required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Can not find Resource to update",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/jobmanager/resources/status/{job_uuid}": {
            "get": {
                "description": "get resource status by job uuid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Get resource status by job UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job UUID",
                        "name": "job_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Resource"
                        }
                    },
                    "400": {
                        "description": "Job UUID is required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Can not find Job by UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Condition": {
            "type": "object",
            "required": [
                "lastTransitionTime",
                "message",
                "reason",
                "status",
                "type"
            ],
            "properties": {
                "lastTransitionTime": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "observedGeneration": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.ConditionStatus"
                },
                "type": {
                    "$ref": "#/definitions/models.ResourceState"
                }
            }
        },
        "models.ConditionStatus": {
            "type": "string",
            "enum": [
                "True",
                "False",
                "Unknown"
            ],
            "x-enum-varnames": [
                "ConditionTrue",
                "ConditionFalse",
                "ConditionUnknown"
            ]
        },
        "models.Content": {
            "type": "object",
            "required": [
                "name",
                "yaml"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "yaml": {
                    "type": "string"
                }
            }
        },
        "models.Instruction": {
            "type": "object",
            "properties": {
                "componentName": {
                    "type": "string"
                },
                "contents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Content"
                    }
                },
                "id": {
                    "type": "string"
                },
                "job_id": {
                    "description": "Excluded from YAML",
                    "type": "string"
                },
                "policies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Policy"
                    }
                },
                "requirements": {
                    "$ref": "#/definitions/models.Requirement"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.Job": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "instruction": {
                    "$ref": "#/definitions/models.Instruction"
                },
                "job_group_id": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "orchestrator": {
                    "$ref": "#/definitions/models.OrchestratorType"
                },
                "owner_id": {
                    "type": "string"
                },
                "resource": {
                    "$ref": "#/definitions/models.Resource"
                },
                "state": {
                    "$ref": "#/definitions/models.JobState"
                },
                "sub_type": {
                    "$ref": "#/definitions/models.RemediationType"
                },
                "targets": {
                    "$ref": "#/definitions/models.Target"
                },
                "type": {
                    "$ref": "#/definitions/models.JobType"
                }
            }
        },
...
        "models.JobState": {
            "type": "string",
            "enum": [
                "Created",
                "Progressing",
                "Finished",
                "Degraded"
            ],
            "x-enum-varnames": [
                "Created",
                "Progressing",
                "Finished",
                "Degraded"
            ]
        },
        "models.JobType": {
            "type": "string",
            "enum": [
                "CreateDeployment",
                "DeleteDeployment",
                "UpdateDeployment"
            ],
            "x-enum-varnames": [
                "CreateDeployment",
                "DeleteDeployment",
                "UpdateDeployment"
            ]
        },
        "models.OrchestratorType": {
            "type": "string",
            "enum": [
                "ocm",
                "nuvla",
                ""
            ],
            "x-enum-varnames": [
                "OCM",
                "Nuvla",
                "None"
            ]
        },
        "models.Policy": {
            "type": "object",
            "properties": {
                "component": {
                    "type": "string"
                },
                "fromTemplate": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "remediation": {
                    "type": "string"
                },
                "spec": {
                    "$ref": "#/definitions/models.PolicySpec"
                },
                "variables": {
                    "$ref": "#/definitions/models.PolicyVariables"
                }
            }
        },
        "models.PolicySpec": {
            "type": "object",
            "properties": {
                "expr": {
                    "type": "string"
                },
                "thresholds": {
                    "$ref": "#/definitions/models.Thresholds"
                }
            }
        },
        "models.PolicyVariables": {
            "type": "object",
            "properties": {
                "compssTask": {
                    "type": "string"
                },
                "thresholdTimeSeconds": {
                    "type": "integer"
                }
            }
        },
        "models.Remediation": {
            "type": "object",
            "required": [
                "remediationStatus",
                "remediationType"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "remediationStatus": {
                    "default": "Pending",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.RemediationStatus"
                        }
                    ]
                },
                "remediationTarget": {
                    "$ref": "#/definitions/models.RemediationTarget"
                },
                "remediationType": {
                    "$ref": "#/definitions/models.RemediationType"
                },
                "resource_id": {
                    "type": "string"
                }
            }
        },
        "models.RemediationStatus": {
            "type": "string",
            "enum": [
                "Pending",
                "Failed",
                "Remediated"
            ],
            "x-enum-varnames": [
                "Pending",
                "Failed",
                "Remediated"
            ]
        },
        "models.RemediationTarget": {
            "type": "object",
            "properties": {
                "command": {
                    "type": "string"
                },
                "container": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "node": {
                    "type": "string"
                },
                "pod": {
                    "type": "string"
                },
                "pod_uid": {
                    "type": "string"
                },
                "remediation_id": {
                    "type": "string"
                }
            }
        },
        "models.RemediationType": {
            "type": "string",
            "enum": [
                "scale-up",
                "scale-down",
                "scale-out",
                "scale-in",
                "patch",
                "reallocate",
                "replace",
                "secure"
            ],
            "x-enum-varnames": [
                "ScaleUp",
                "ScaleDown",
                "ScaleOut",
                "ScaleIn",
                "Patch",
                "Reallocate",
                "Replace",
                "Secure"
            ]
        },
        "models.Requirement": {
            "type": "object",
            "properties": {
                "architecture": {
                    "type": "string"
                },
                "cpu": {
                    "type": "string"
                },
                "devices": {
                    "type": "string"
                },
                "memory": {
                    "type": "string"
                }
            }
        },
        "models.Resource": {
            "type": "object",
            "properties": {
                "conditions": {
                    "description": "OriginalResourceName string      `gorm:\"type:text\" json:\"-\" validate:\"omitempty\"`",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Condition"
                    }
                },
                "job_id": {
                    "type": "string"
                },
                "remediations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Remediation"
                    }
                },
                "resource_name": {
                    "type": "string"
                },
                "resource_uuid": {
                    "description": "we set this field manually",
                    "type": "string"
                }
            }
        },
        "models.ResourceState": {
            "type": "string",
            "enum": [
                "Progressing",
                "Applied",
                "Available",
                "Degraded"
            ],
            "x-enum-varnames": [
                "ResourceProgressing",
                "ResourceApplied",
                "ResourceAvailable",
                "ResourceDegraded"
            ]
        },
        "models.Target": {
            "type": "object",
            "required": [
                "cluster_name",
                "orchestrator"
            ],
            "properties": {
                "cluster_name": {
                    "type": "string"
                },
                "node_name": {
                    "type": "string"
                },
                "orchestrator": {
                    "$ref": "#/definitions/models.OrchestratorType"
                }
            }
        },
        "models.Thresholds": {
            "type": "object",
            "properties": {
                "critical": {
                    "type": "integer"
                },
                "warning": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "\"Type 'Bearer TOKEN' to correctly set the API Key\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}